requirements.txt
numpy~=1.26.4
setuptools~=75.8.0
cython~=3.1.0
scikit-learn~=1.6.1
trimesh~=4.6.8
scikit-image~=0.24.0
scipy~=1.13.1
chumpy~=0.70
opencv-python~=4.11.0.86
ipdb~=0.13.13
matplotlib~=3.9.4
imageio~=2.37.0
open3d~=0.19.0
tqdm~=4.67.1
pytorch3d~=0.7.7
yaml~=0.2.5
pyyaml~=6.0.2
pointops~=0.0.0
kaolin~=0.17.0
smplx~=0.1.28